Update README.md
the description will literally be longer than the change made here... well, not here, in the code.
I did change stuff here... well not "change" this is the original description,  which is a change I suppose from no description, but not like there was an original one, then I changed it.  I just created the original one.
the change I made in the readme.txt file was a remark on how cool this was.  And I don't want to get hung up on "change" but it was really more of an "addition", which does constitute a change.  But when I think of change I don't think.... aaahhhhhhggg!!